+++ SIL.Windows.Forms/Reporting/ExceptionReportingDialog.cs
@@ -5,11 +5,11 @@ using System.Drawing;
 using System.Text;
 using System.Threading;
 using System.Windows.Forms;
+using L10NSharp;
 using SIL.Email;
 using SIL.Reporting;
-using System.Runtime.InteropServices;
-using SIL.PlatformUtilities;
 using SIL.Windows.Forms.Miscellaneous;
+using static System.String;
 
 namespace SIL.Windows.Forms.Reporting
 {
@@ -34,31 +34,34 @@ namespace SIL.Windows.Forms.Reporting
 				ThreadId = threadId;
 			}
 
-			public string Message;
-			public string MessageBeforeStack;
-			public Exception Error;
-			public Form OwningForm;
-			public StackTrace StackTrace;
-			public int ThreadId;
+			public readonly string Message;
+			public readonly string MessageBeforeStack;
+			public readonly Exception Error;
+			public readonly Form OwningForm;
+			public readonly StackTrace StackTrace;
+			public readonly int ThreadId;
 		}
 		#endregion
 
 		#region Member variables
 
-		private Label label3;
+		private Label _detailsForDevelopers;
 		private TextBox _details;
-		private TextBox _pleaseHelpText;
-		private TextBox m_reproduce;
+		private Label _lblPleaseHelp;
+		private TextBox _reproduce;
 
-		private bool _isLethal;
+		private readonly bool _isLethal;
 
 		private Button _sendAndCloseButton;
-		private TextBox _notificationText;
-		private TextBox textBox1;
+		private Label _notificationText;
+		private Label _lblStepsToReproduce;
 		private ComboBox _methodCombo;
 		private Button _privacyNoticeButton;
 		private Label _emailAddress;
 		private static bool s_doIgnoreReport;
+		private TableLayoutPanel tableLayoutPanel1;
+		private System.ComponentModel.IContainer components;
+
 		/// <summary>
 		/// Stack with exception data.
 		/// </summary>
@@ -75,7 +78,8 @@ namespace SIL.Windows.Forms.Reporting
 		/// the exception (similar to the behavior we already have when we get an exception on
 		/// the UI thread while displaying the exception dialog). Otherwise we display the
 		/// exception dialog, appending the messages from the exception data stack.</remarks>
-		private static Stack<ExceptionReportingData> s_reportDataStack = new Stack<ExceptionReportingData>();
+		private static readonly Stack<ExceptionReportingData> s_reportDataStack =
+			new Stack<ExceptionReportingData>();
 
 		#endregion
 
@@ -95,7 +99,8 @@ namespace SIL.Windows.Forms.Reporting
 		{
 			if (IsDisposed)
 			{
-				throw new ObjectDisposedException(String.Format("'{0}' in use after being disposed.", GetType().Name));
+				throw new ObjectDisposedException(
+					$"'{GetType().Name}' in use after being disposed.");
 			}
 		}
 
@@ -149,141 +154,226 @@ namespace SIL.Windows.Forms.Reporting
 		/// </summary>
 		private void InitializeComponent()
 		{
+			this.components = new System.ComponentModel.Container();
 			System.ComponentModel.ComponentResourceManager resources = new System.ComponentModel.ComponentResourceManager(typeof(ExceptionReportingDialog));
-			this.m_reproduce = new System.Windows.Forms.TextBox();
-			this.label3 = new System.Windows.Forms.Label();
+			this._reproduce = new System.Windows.Forms.TextBox();
+			this._detailsForDevelopers = new System.Windows.Forms.Label();
 			this._details = new System.Windows.Forms.TextBox();
 			this._sendAndCloseButton = new System.Windows.Forms.Button();
-			this._pleaseHelpText = new System.Windows.Forms.TextBox();
-			this._notificationText = new System.Windows.Forms.TextBox();
-			this.textBox1 = new System.Windows.Forms.TextBox();
+			this._lblPleaseHelp = new System.Windows.Forms.Label();
+			this._notificationText = new System.Windows.Forms.Label();
+			this._lblStepsToReproduce = new System.Windows.Forms.Label();
 			this._methodCombo = new System.Windows.Forms.ComboBox();
 			this._privacyNoticeButton = new System.Windows.Forms.Button();
 			this._emailAddress = new System.Windows.Forms.Label();
+			this.tableLayoutPanel1 = new System.Windows.Forms.TableLayoutPanel();
+			this.tableLayoutPanel1.SuspendLayout();
 			this.SuspendLayout();
-			//
+			// 
 			// m_reproduce
-			//
-			this.m_reproduce.AcceptsReturn = true;
-			this.m_reproduce.AcceptsTab = true;
-			resources.ApplyResources(this.m_reproduce, "m_reproduce");
-			this.m_reproduce.Name = "m_reproduce";
-			//
-			// label3
-			//
-			resources.ApplyResources(this.label3, "label3");
-			this.label3.Name = "label3";
-			//
+			// 
+			this._reproduce.AcceptsReturn = true;
+			this._reproduce.AcceptsTab = true;
+			resources.ApplyResources(this._reproduce, "m_reproduce");
+			this.tableLayoutPanel1.SetColumnSpan(this._reproduce, 2);
+			this._reproduce.Name = "m_reproduce";
+			// 
+			// _detailsForDevelopers
+			// 
+			resources.ApplyResources(this._detailsForDevelopers, "_detailsForDevelopers");
+			this.tableLayoutPanel1.SetColumnSpan(this._detailsForDevelopers, 2);
+			this._detailsForDevelopers.Name = "_detailsForDevelopers";
+			// 
 			// _details
-			//
+			// 
 			resources.ApplyResources(this._details, "_details");
 			this._details.BackColor = System.Drawing.SystemColors.ControlLightLight;
+			this.tableLayoutPanel1.SetColumnSpan(this._details, 2);
 			this._details.Name = "_details";
 			this._details.ReadOnly = true;
-			//
+			// 
 			// _sendAndCloseButton
-			//
+			// 
 			resources.ApplyResources(this._sendAndCloseButton, "_sendAndCloseButton");
 			this._sendAndCloseButton.DialogResult = System.Windows.Forms.DialogResult.Cancel;
 			this._sendAndCloseButton.Name = "_sendAndCloseButton";
 			this._sendAndCloseButton.Click += new System.EventHandler(this.btnClose_Click);
-			//
-			// _pleaseHelpText
-			//
-			resources.ApplyResources(this._pleaseHelpText, "_pleaseHelpText");
-			this._pleaseHelpText.BackColor = System.Drawing.Color.FromArgb(((int)(((byte)(192)))), ((int)(((byte)(255)))), ((int)(((byte)(192)))));
-			this._pleaseHelpText.BorderStyle = System.Windows.Forms.BorderStyle.None;
-			this._pleaseHelpText.ForeColor = System.Drawing.Color.Black;
-			this._pleaseHelpText.Name = "_pleaseHelpText";
-			this._pleaseHelpText.ReadOnly = true;
-			//
+			// 
+			// _lblPleaseHelp
+			// 
+			resources.ApplyResources(this._lblPleaseHelp, "_lblPleaseHelp");
+			this._lblPleaseHelp.BackColor = System.Drawing.Color.FromArgb(((int)(((byte)(192)))), ((int)(((byte)(255)))), ((int)(((byte)(192)))));
+			this.tableLayoutPanel1.SetColumnSpan(this._lblPleaseHelp, 2);
+			this._lblPleaseHelp.ForeColor = System.Drawing.Color.Black;
+			this._lblPleaseHelp.Name = "_lblPleaseHelp";
+			// 
 			// _notificationText
-			//
+			// 
 			resources.ApplyResources(this._notificationText, "_notificationText");
 			this._notificationText.BackColor = System.Drawing.Color.FromArgb(((int)(((byte)(192)))), ((int)(((byte)(255)))), ((int)(((byte)(192)))));
-			this._notificationText.BorderStyle = System.Windows.Forms.BorderStyle.None;
+			this.tableLayoutPanel1.SetColumnSpan(this._notificationText, 2);
 			this._notificationText.ForeColor = System.Drawing.Color.Black;
 			this._notificationText.Name = "_notificationText";
-			this._notificationText.ReadOnly = true;
-			//
-			// textBox1
-			//
-			resources.ApplyResources(this.textBox1, "textBox1");
-			this.textBox1.BackColor = System.Drawing.Color.FromArgb(((int)(((byte)(192)))), ((int)(((byte)(255)))), ((int)(((byte)(192)))));
-			this.textBox1.BorderStyle = System.Windows.Forms.BorderStyle.None;
-			this.textBox1.ForeColor = System.Drawing.Color.Black;
-			this.textBox1.Name = "textBox1";
-			this.textBox1.ReadOnly = true;
-			//
+			// 
+			// _lblStepsToReproduce
+			// 
+			resources.ApplyResources(this._lblStepsToReproduce, "_lblStepsToReproduce");
+			this._lblStepsToReproduce.BackColor = System.Drawing.Color.FromArgb(((int)(((byte)(192)))), ((int)(((byte)(255)))), ((int)(((byte)(192)))));
+			this.tableLayoutPanel1.SetColumnSpan(this._lblStepsToReproduce, 2);
+			this._lblStepsToReproduce.ForeColor = System.Drawing.Color.Black;
+			this._lblStepsToReproduce.Name = "_lblStepsToReproduce";
+			// 
 			// _methodCombo
-			//
+			// 
 			resources.ApplyResources(this._methodCombo, "_methodCombo");
 			this._methodCombo.DropDownStyle = System.Windows.Forms.ComboBoxStyle.DropDownList;
 			this._methodCombo.FormattingEnabled = true;
 			this._methodCombo.Name = "_methodCombo";
 			this._methodCombo.SelectedIndexChanged += new System.EventHandler(this._methodCombo_SelectedIndexChanged);
-			//
+			// 
 			// _privacyNoticeButton
-			//
+			// 
 			resources.ApplyResources(this._privacyNoticeButton, "_privacyNoticeButton");
 			this._privacyNoticeButton.Image = global::SIL.Windows.Forms.Properties.Resources.spy16x16;
 			this._privacyNoticeButton.Name = "_privacyNoticeButton";
 			this._privacyNoticeButton.UseVisualStyleBackColor = true;
 			this._privacyNoticeButton.Click += new System.EventHandler(this._privacyNoticeButton_Click);
-			//
+			// 
 			// _emailAddress
-			//
-			resources.ApplyResources(this._emailAddress, "_emailAddress");
+			// 
 			this._emailAddress.ForeColor = System.Drawing.Color.DimGray;
+			resources.ApplyResources(this._emailAddress, "_emailAddress");
 			this._emailAddress.Name = "_emailAddress";
-			//
+			// 
+			// tableLayoutPanel1
+			// 
+			resources.ApplyResources(this.tableLayoutPanel1, "tableLayoutPanel1");
+			this.tableLayoutPanel1.Controls.Add(this._notificationText, 0, 0);
+			this.tableLayoutPanel1.Controls.Add(this._sendAndCloseButton, 1, 7);
+			this.tableLayoutPanel1.Controls.Add(this._privacyNoticeButton, 0, 7);
+			this.tableLayoutPanel1.Controls.Add(this._emailAddress, 0, 6);
+			this.tableLayoutPanel1.Controls.Add(this._methodCombo, 1, 6);
+			this.tableLayoutPanel1.Controls.Add(this._lblPleaseHelp, 0, 1);
+			this.tableLayoutPanel1.Controls.Add(this._lblStepsToReproduce, 0, 2);
+			this.tableLayoutPanel1.Controls.Add(this._reproduce, 0, 3);
+			this.tableLayoutPanel1.Controls.Add(this._details, 0, 5);
+			this.tableLayoutPanel1.Controls.Add(this._detailsForDevelopers, 0, 4);
+			this.tableLayoutPanel1.Name = "tableLayoutPanel1";
+			// 
 			// ExceptionReportingDialog
-			//
+			// 
 			this.AcceptButton = this._sendAndCloseButton;
 			resources.ApplyResources(this, "$this");
 			this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Dpi;
 			this.BackColor = System.Drawing.Color.FromArgb(((int)(((byte)(192)))), ((int)(((byte)(255)))), ((int)(((byte)(192)))));
-			this.Controls.Add(this._emailAddress);
-			this.Controls.Add(this._privacyNoticeButton);
-			this.Controls.Add(this._methodCombo);
-			this.Controls.Add(this.textBox1);
-			this.Controls.Add(this.m_reproduce);
-			this.Controls.Add(this._notificationText);
-			this.Controls.Add(this._pleaseHelpText);
-			this.Controls.Add(this._details);
-			this.Controls.Add(this.label3);
-			this.Controls.Add(this._sendAndCloseButton);
+			this.Controls.Add(this.tableLayoutPanel1);
 			this.KeyPreview = true;
 			this.MaximizeBox = false;
 			this.MinimizeBox = false;
 			this.Name = "ExceptionReportingDialog";
-			this.ControlBox = true;
-			this.ShowIcon = false; // Showing the Control box ("X") also shows a default icon.
+			this.ShowIcon = false;
+			this.SizeGripStyle = System.Windows.Forms.SizeGripStyle.Hide;
 			this.TopMost = true;
 			this.KeyPress += new System.Windows.Forms.KeyPressEventHandler(this.ExceptionReportingDialog_KeyPress);
+			this.tableLayoutPanel1.ResumeLayout(false);
+			this.tableLayoutPanel1.PerformLayout();
 			this.ResumeLayout(false);
-			this.PerformLayout();
 
 		}
 
 		private void SetupMethodCombo()
 		{
 			_methodCombo.Items.Clear();
-			_methodCombo.Items.Add(new ReportingMethod("Send using my email program", "&Email", "mapiWithPopup", SendViaEmail));
-			_methodCombo.Items.Add(new ReportingMethod("Copy to clipboard", "&Copy", "clipboard", PutOnClipboard));
+			string reportingMethod, closeButtonLabelLethal, closeButtonLabelNonLethal;
+			try
+			{
+				reportingMethod = LocalizationManager.GetString(
+					"ExceptionReportingDialog.ReportingMethod.Email",
+					"Send using my email program");
+			}
+			catch (Exception e)
+			{
+				try { Logger.WriteError(e); } catch { /* We tried. */ }
+				reportingMethod = "Send using my email program";
+			}
+			try
+			{
+				closeButtonLabelLethal = LocalizationManager.GetString(
+					"ExceptionReportingDialog.ReportingMethod.EmailCloseButtonLabelLethal",
+					"&Email and Exit", "Ampersand is optional, to indicate accelerator key");
+			}
+			catch (Exception e)
+			{
+				try { Logger.WriteError(e); } catch { /* We tried. */ }
+				closeButtonLabelLethal = "&Email and Exit";
+			}
+			try
+			{
+				closeButtonLabelNonLethal = LocalizationManager.GetString(
+					"ExceptionReportingDialog.ReportingMethod.EmailCloseButtonLabel",
+					"&Email", "Ampersand is optional, to indicate accelerator key");
+			}
+			catch (Exception e)
+			{
+				try { Logger.WriteError(e); } catch { /* We tried. */ }
+				closeButtonLabelNonLethal = "&Email";
+			}
+			_methodCombo.Items.Add(new ReportingMethod(
+				reportingMethod, closeButtonLabelLethal, closeButtonLabelNonLethal,
+				"mapiWithPopup", SendViaEmail));
+			
+			try
+			{
+				reportingMethod = LocalizationManager.GetString(
+					"ExceptionReportingDialog.ReportingMethod.CopyToClipboard",
+					"Copy to clipboard");
+			}
+			catch (Exception e)
+			{
+				try { Logger.WriteError(e); } catch { /* We tried. */ }
+				reportingMethod = "Copy to clipboard";
+			}
+			try
+			{
+				closeButtonLabelLethal = LocalizationManager.GetString(
+					"ExceptionReportingDialog.ReportingMethod.CopyCloseButtonLabelLethal",
+					"&Copy and Exit", "Ampersand is optional, to indicate accelerator key");
+			}
+			catch (Exception e)
+			{
+				try { Logger.WriteError(e); } catch { /* We tried. */ }
+				closeButtonLabelLethal = "&Copy and Exit";
+			}
+			try
+			{
+				closeButtonLabelNonLethal = LocalizationManager.GetString(
+					"ExceptionReportingDialog.ReportingMethod.CopyCloseButtonLabel",
+					"&Copy", "Ampersand is optional, to indicate accelerator key");
+			}
+			catch (Exception e)
+			{
+				try { Logger.WriteError(e); } catch { /* We tried. */ }
+				closeButtonLabelNonLethal = "&Copy";
+			}
+			_methodCombo.Items.Add(new ReportingMethod(
+				reportingMethod, closeButtonLabelLethal, closeButtonLabelNonLethal,
+				"clipboard", PutOnClipboard));
 		}
 
-		class ReportingMethod
+		private class ReportingMethod
 		{
 			private readonly string _label;
-			public readonly string CloseButtonLabel;
+			public readonly string CloseButtonLabelLethal;
+			public readonly string CloseButtonLabelNonLethal;
 			public readonly string Id;
 			public readonly Func<bool> Method;
 
-			public ReportingMethod(string label, string closeButtonLabel, string id, Func<bool> method)
+			public ReportingMethod(string label, string closeButtonLabelLethal,
+				string closeButtonLabelNonLethal, string id, Func<bool> method)
 			{
 				_label = label;
-				CloseButtonLabel = closeButtonLabel;
+				CloseButtonLabelLethal = closeButtonLabelLethal;
+				CloseButtonLabelNonLethal = closeButtonLabelNonLethal;
 				Id = id;
 				Method = method;
 			}
@@ -333,13 +423,11 @@ namespace SIL.Windows.Forms.Reporting
 					s_reportDataStack.Push(new ExceptionReportingData(null, null, error,
 						null, parent, Thread.CurrentThread.ManagedThreadId));
 				}
-				return;            // ignore message if we are showing from a previous error
+				return; // ignore message if we are showing from a previous error
 			}
 
 			using (ExceptionReportingDialog dlg = new ExceptionReportingDialog(isLethal))
-			{
 				dlg.Report(error, parent);
-			}
 		}
 
 		internal static void ReportMessage(string message, StackTrace stack, bool isLethal)
@@ -351,13 +439,11 @@ namespace SIL.Windows.Forms.Reporting
 					s_reportDataStack.Push(new ExceptionReportingData(message, null, null,
 						stack, null, Thread.CurrentThread.ManagedThreadId));
 				}
-				return;            // ignore message if we are showing from a previous error
+				return; // ignore message if we are showing from a previous error
 			}
 
 			using (ExceptionReportingDialog dlg = new ExceptionReportingDialog(isLethal))
-			{
-				dlg.Report(message, string.Empty, stack, null);
-			}
+				dlg.Report(message, Empty, stack, null);
 		}
 
 		internal static void ReportMessage(string message, Exception error, bool isLethal)
@@ -369,7 +455,7 @@ namespace SIL.Windows.Forms.Reporting
 					s_reportDataStack.Push(new ExceptionReportingData(message, null, error,
 						null, null, Thread.CurrentThread.ManagedThreadId));
 				}
-				return;            // ignore message if we are showing from a previous error
+				return; // ignore message if we are showing from a previous error
 			}
 
 			using (ExceptionReportingDialog dlg = new ExceptionReportingDialog(isLethal))
@@ -380,7 +466,7 @@ namespace SIL.Windows.Forms.Reporting
 
 		protected void GatherData()
 		{
-			_details.Text += Environment.NewLine + "To Reproduce: " + m_reproduce.Text + Environment.NewLine;
+			_details.Text += Environment.NewLine + "To Reproduce: " + _reproduce.Text + Environment.NewLine;
 		}
 
 		public void Report(Exception error, Form owningForm)
@@ -447,7 +533,7 @@ namespace SIL.Windows.Forms.Reporting
 
 			PrepareDialog();
 
-			if(!string.IsNullOrEmpty(reportingData.Message))
+			if (!IsNullOrEmpty(reportingData.Message))
 				_notificationText.Text = reportingData.Message;
 
 			var bldr = new StringBuilder();
@@ -470,7 +556,7 @@ namespace SIL.Windows.Forms.Reporting
 				}
 				catch (Exception err)
 				{
-					//We have more than one report of dieing while logging an exception.
+					//We have more than one report of dying while logging an exception.
 					_details.Text += "****Could not write to log (" + err.Message + ")" + Environment.NewLine;
 					_details.Text += "Was trying to log the exception: " + error.Message + Environment.NewLine;
 					_details.Text += "Recent events:" + Environment.NewLine;
@@ -485,7 +571,7 @@ namespace SIL.Windows.Forms.Reporting
 				}
 				catch (Exception err)
 				{
-					//We have more than one report of dieing while logging an exception.
+					//We have more than one report of dying while logging an exception.
 					_details.Text += "****Could not write to log (" + err.Message + ")" + Environment.NewLine;
 				}
 			}
@@ -497,7 +583,7 @@ namespace SIL.Windows.Forms.Reporting
 		{
 			try
 			{
-				if (!string.IsNullOrEmpty(reportingData.Message))
+				if (!IsNullOrEmpty(reportingData.Message))
 					UsageReporter.ReportExceptionString(reportingData.Message);
 				else if (reportingData.Error != null)
 					UsageReporter.ReportException(reportingData.Error);
@@ -539,13 +625,13 @@ namespace SIL.Windows.Forms.Reporting
 
 		private static Exception FormatMessage(StringBuilder bldr, ExceptionReportingData data)
 		{
-			if (!string.IsNullOrEmpty(data.Message))
+			if (!IsNullOrEmpty(data.Message))
 			{
 				bldr.Append("Message (not an exception): ");
 				bldr.AppendLine(data.Message);
 				bldr.AppendLine();
 			}
-			if (!string.IsNullOrEmpty(data.MessageBeforeStack))
+			if (!IsNullOrEmpty(data.MessageBeforeStack))
 			{
 				bldr.AppendLine(data.MessageBeforeStack);
 			}
@@ -554,8 +640,8 @@ namespace SIL.Windows.Forms.Reporting
 			{
 				Exception innerMostException = null;
 				bldr.Append(ErrorReport.GetHiearchicalExceptionInfo(data.Error, ref innerMostException));
-				//if the exception had inner exceptions, show the inner-most exception first, since that is usually the one
-				//we want the developer to read.
+				// If the exception had inner exceptions, show the inner-most exception first,
+				// since that is likely to be the one most useful for the developer to read first.
 				if (innerMostException != null)
 				{
 					var oldText = bldr.ToString();
@@ -626,13 +712,28 @@ namespace SIL.Windows.Forms.Reporting
 			if (!_isLethal)
 			{
 				BackColor = Color.FromArgb(255, 255, 192); //yellow
-				_notificationText.Text = "Take Courage. It'll work out.";
+				try
+				{
+					_notificationText.Text = LocalizationManager.GetString(
+						"ExceptionReportingDialog.NotificationTextNonLethal",
+						"Take Courage. It'll work out.",
+						"For more formal cultures, something less cheeky might be better. In most " +
+						"cases, this will be replaced by a more useful message, but there are " +
+						"situations where a user could see this.");
+				}
+				catch (Exception e)
+				{
+					try { Logger.WriteError(e); } catch { /* We tried. */ }
+				}
+
 				_notificationText.BackColor = BackColor;
-				_pleaseHelpText.BackColor = BackColor;
-				textBox1.BackColor = BackColor;
+				_lblPleaseHelp.BackColor = BackColor;
+				_lblStepsToReproduce.BackColor = BackColor;
 			}
 
 			SetupCloseButtonText();
+
+			SetLocalizedText();
 		}
 
 		private void ShowReportDialogIfAppropriate(Form owningForm)
@@ -650,8 +751,7 @@ namespace SIL.Windows.Forms.Reporting
 
 
 		}
-
-
+		
 		/// ------------------------------------------------------------------------------------
 		/// <summary>
 		///
@@ -687,7 +787,24 @@ namespace SIL.Windows.Forms.Reporting
 		{
 			if (ErrorReport.EmailAddress != null)
 			{
-				_details.Text = String.Format("Please e-mail this to {0} {1}", ErrorReport.EmailAddress, _details.Text);
+				string localizedEmailInstructions = null;
+				try
+				{
+					if (LocalizationManager.UILanguageId != "en")
+					{
+						localizedEmailInstructions = LocalizationManager.GetString(
+							"ExceptionReportingDialog.EmailInstructions",
+							"(Please e-mail this to)");
+						if (localizedEmailInstructions == "(Please e-mail this to)")
+							localizedEmailInstructions = null;
+					}
+				}
+				catch (Exception e)
+				{
+					try { Logger.WriteError(e); } catch { /* We tried. */ }
+				}
+				_details.Text = $"Please e-mail this to {localizedEmailInstructions} {ErrorReport.EmailAddress}" +
+					$"{Environment.NewLine}{_details.Text}";
 			}
 			try
 			{
@@ -793,7 +910,15 @@ namespace SIL.Windows.Forms.Reporting
 		{
 			if (e.KeyCode == Keys.ShiftKey && Visible)
 			{
-				_sendAndCloseButton.Text = "Continue";
+				try
+				{
+					_sendAndCloseButton.Text = LocalizationManager.GetString(
+						"ExceptionReportingDialog.ContinueButtonLabel", "Continue");
+				}
+				catch (Exception ex)
+				{
+					try { Logger.WriteError(ex); } catch { /* We tried. */ }
+				}
 			}
 			base.OnKeyDown(e);
 		}
@@ -813,55 +938,225 @@ namespace SIL.Windows.Forms.Reporting
 			base.OnKeyUp(e);
 		}
 
-		private void OnJustExit_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
-		{
-			CloseUp();
-		}
-
 		private void _methodCombo_SelectedIndexChanged(object sender, EventArgs e)
 		{
 			SetupCloseButtonText();
 		}
 
-		private void SetupCloseButtonText()
+		private void SetupCloseButtonText() =>
+			_sendAndCloseButton.Text = _isLethal ? SelectedMethod.CloseButtonLabelLethal :
+				SelectedMethod.CloseButtonLabelNonLethal;
+
+		private void SetLocalizedText()
 		{
-			_sendAndCloseButton.Text = SelectedMethod.CloseButtonLabel;
-			if (!_isLethal)
+			try
 			{
-				// _dontSendEmailLink.Text = "Don't Send Email";
+				if (LocalizationManager.UILanguageId == "en")
+					return;
 			}
-			else
+			catch (Exception e)
 			{
-				_sendAndCloseButton.Text += " and Exit";
+				try { Logger.WriteError(e); } catch { /* We tried. */ }
+				// If we couldn't get the current UILanguageId without crashing, we don't have much
+				// hope of retrieving localized strings. Just use what Designer resources supplied.
+				return;
+			}
+
+			// All of these strings should be kept in sync with the ones in the Designer-generated
+			// We opted to do it this way instead of using the LocalizationExtender so we could
+			// ensure that applying localizations to the error reporting form would not itself
+			// result in a fatal error.
+			try
+			{
+				Text = LocalizationManager.GetString("ExceptionReportingDialog.WindowTitle",
+					"Error");
+			}
+			catch (Exception e)
+			{
+				try { Logger.WriteError(e); } catch { /* We tried. */ }
+			}
+
+			try
+			{
+				_detailsForDevelopers.Text = LocalizationManager.GetString(
+					"ExceptionReportingDialog._detailsForDevelopers",
+					"Details for the developers:");
+			}
+			catch (Exception e)
+			{
+				try { Logger.WriteError(e); } catch { /* We tried. */ }
+			}
+
+			try
+			{
+				_lblPleaseHelp.Text = LocalizationManager.GetString(
+					"ExceptionReportingDialog._lblPleaseHelp",
+					"To help us fix the problem, we need to gather information on what went wrong.");
+			}
+			catch (Exception e)
+			{
+				try { Logger.WriteError(e); } catch { /* We tried. */ }
+			}
+
+			if (_isLethal)
+			{
+				try
+				{
+					_notificationText.Text = LocalizationManager.GetString(
+						"ExceptionReportingDialog.NotificationTextLethal",
+						"Well, this is embarrassing.",
+						"For more formal cultures, something less cheeky might be better. In most " +
+						"cases, this will be replaced by a more useful message, but there are " +
+						"situations where a user could see this.");
+				}
+				catch (Exception e)
+				{
+					try { Logger.WriteError(e); } catch { /* We tried. */ }
+				}
+			}
+
+			try
+			{
+				_lblStepsToReproduce.Text = LocalizationManager.GetString(
+					"ExceptionReportingDialog._lblStepsToReproduce",
+					"Can you list the steps to take to make this happen again, or tell us anything that might be helpful?");
+			}
+			catch (Exception e)
+			{
+				try { Logger.WriteError(e); } catch { /* We tried. */ }
+			}
+
+			try
+			{
+				_privacyNoticeButton.Text = LocalizationManager.GetString(
+					"ExceptionReportingDialog._privacyNoticeButton",
+					"Privacy Notice...");
+			}
+			catch (Exception e)
+			{
+				try { Logger.WriteError(e); } catch { /* We tried. */ }
 			}
 		}
 
 		private ReportingMethod SelectedMethod
 		{
-			get { return ((ReportingMethod) _methodCombo.SelectedItem); }
-			set { _methodCombo.SelectedItem = value; }
+			get => ((ReportingMethod) _methodCombo.SelectedItem);
+			set => _methodCombo.SelectedItem = value;
 		}
 
-		public static string PrivacyNotice = @"If you don't care who reads your bug report, you can skip this notice.
+		/// <summary>
+		/// Although this is public and can be set externally for an application that has a custom
+		/// policy (i.e., does not use Jira), be aware that if you want it to be localized, you
+		/// will need to reset it if the UI locale changes.
+		/// </summary>
+		public static string PrivacyNotice;
 
-When you submit a crash report or other issue, the contents of your email go in our issue tracking system, ""jira"", which is available via the web at https://jira.sil.org/issues. This is the normal way to handle issues in an open-source project.
+		private static string LocalizedPrivacyNotice => PrivacyNotice ?? GetDefaultPrivacyNotice();
 
-Our issue-tracking system is not searchable by those without an account. Therefore, someone searching via Google will not find your bug reports.
+		private static string GetDefaultPrivacyNotice()
+		{
+			var bldr = new StringBuilder();
+			string nextPara;
+			try
+			{
+				nextPara = LocalizationManager.GetString("ExceptionReportingDialog.Privacy.Para1",
+					"If you don't care who reads your bug report, you can disregard this notice.");
+			}
+			catch (Exception e)
+			{
+				try { Logger.WriteError(e); } catch { /* We tried. */ }
 
-However, anyone can make an account and then read what you sent us. So if you have something private to say, please send it to one of the developers privately with a note that you don't want the issue in our issue tracking system. If need be, we'll make some kind of sanitized place-holder for your issue so that we don't lose it.
-";
+				nextPara = "If you don't care who reads your bug report, you can disregard this " +
+					"notice.";
+			}
+			bldr.Append(nextPara);
+			bldr.Append(Environment.NewLine);
+			bldr.Append(Environment.NewLine);
+
+			// REVIEW: Is the last line (which I have put in parentheses) true and/or helpful?
+			try
+			{
+				nextPara = LocalizationManager.GetString("ExceptionReportingDialog.Privacy.Para2",
+					"When you submit a crash report or other issue, it goes into our issue " +
+					"tracking system, \"Jira\", which is available via the web at " +
+					"https://jira.sil.org/issues." +
+					"(This is a normal way to handle issues in an open-source project.)");
+			}
+			catch (Exception e)
+			{
+				try { Logger.WriteError(e); } catch { /* We tried. */ }
+
+				nextPara = "When you submit a crash report or other issue, it goes into our issue " +
+					"tracking system, \"Jira\", which is available via the web at " +
+					"https://jira.sil.org/issues." +
+					"(This is a normal way to handle issues in an open-source project.)";
+			}
+			bldr.Append(nextPara);
+			bldr.Append(Environment.NewLine);
+			bldr.Append(Environment.NewLine);
+
+			try
+			{
+				nextPara = LocalizationManager.GetString("ExceptionReportingDialog.Privacy.Para3",
+					"Our issue-tracking system is not searchable by those without an account. " +
+					"Therefore, search engines (like Google) will not find your bug reports.");
+			}
+			catch (Exception e)
+			{
+				try { Logger.WriteError(e); } catch { /* We tried. */ }
+
+				nextPara = "Our issue-tracking system is not searchable by those without an " +
+					"account. Therefore, search engines (like Google) will not find your bug " +
+					"reports.";
+			}
+			bldr.Append(nextPara);
+			bldr.Append(Environment.NewLine);
+			bldr.Append(Environment.NewLine);
+			
+			try
+			{
+				nextPara = LocalizationManager.GetString("ExceptionReportingDialog.Privacy.Para4",
+					"However, anyone can make an account and then read your report. So if you " +
+					"have something private to say, please send it to one of the developers " +
+					"privately with a note that you don't want the issue in our issue tracking " +
+					"system. If necessary, we can make a place-holder for your issue without " +
+					"the private information so the issue can still be tracked and not get lost.");
+			}
+			catch (Exception e)
+			{
+				try { Logger.WriteError(e); } catch { /* We tried. */ }
+
+				nextPara = "However, anyone can make an account and then read your report. So if" +
+					" you have something private to say, please send it to one of the developers" +
+					" privately with a note that you don't want the issue in our issue tracking " +
+					"system. If necessary, we can make a place-holder for your issue without " +
+					"the private information so the issue can still be tracked and not get lost.";
+			}
+			bldr.Append(nextPara);
+
+			return bldr.ToString();
+		}
 
 		private void _privacyNoticeButton_Click(object sender, EventArgs e)
 		{
-			MessageBox.Show(PrivacyNotice, "Privacy Notice");
+			string caption;
+			try
+			{
+				caption = LocalizationManager.GetString(
+					"ExceptionReportingDialog.PrivacyNoticeCaption", "Privacy Notice");
+			}
+			catch (Exception ex)
+			{
+				try { Logger.WriteError(ex); } catch { /* We tried. */ }
+				caption = "Privacy Notice";
+			}
+			MessageBox.Show(LocalizedPrivacyNotice, caption);
 		}
 
 		private void ExceptionReportingDialog_KeyPress(object sender, KeyPressEventArgs e)
 		{
-			if(e.KeyChar== 27)//ESCAPE
-			{
+			if (e.KeyChar== 27)//ESCAPE
 				CloseUp();
-			}
 		}
 	}
 }